nogap-workspace/resume_parser/requirements.txt
pyresparser==1.0.6
spacy==3.7.4
nltk==3.8.1
python-docx==1.1.0
pdfminer.six==20231228
